Fix syri end

+++ wrapper/iga/project/a188.py
@@ -572,7 +572,7 @@ def synal_to_paf(synal_file=None):
     return 0
 
 
-def fix_syri_end(syri_out=None, qry_fa=None, ref_fa=None):
+def fix_syri_end(syri_out=None, qry_fa=None, ref_fa=None, postprocess='F'):
     """
     SyRI fails to find syntenic region at the end of chromosome1. resulting 40 Mb FP unsyntenic regions
     This script is used to fix this problem. The input argument is directory where syri was executed
@@ -590,23 +590,27 @@ def fix_syri_end(syri_out=None, qry_fa=None, ref_fa=None):
     ref_offset = int(mylist[7]) + 1
     ref_chr = mylist[5]
 
-    qry_fadt = SeqIO.to_dict(SeqIO.parse(qry_fa, "fasta"))
-    ref_fadt = SeqIO.to_dict(SeqIO.parse(ref_fa, "fasta"))
-
     qry_tail_fa = "{}_tail.fa".format(qry_fa)
     ref_tail_fa = "{}_tail.fa".format(ref_fa)
-    with open(qry_tail_fa, 'w') as out_qry:
-        out_qry.write(qry_fadt[qry_chr][qry_offset - 1:].format('fasta'))
-    with open(ref_tail_fa, 'w') as out_ref:
-        out_ref.write(ref_fadt[ref_chr][ref_offset - 1:].format('fasta'))
-    syri(qry_tail_fa, ref_tail_fa, submit='T')
-    # formated_SYN = format_syri_offset(qry_offset, ref_offset, "{}.{}.syri.out".format(qry_tail_fa, ref_tail_fa))
-    # curated_SYN_file = syri_out + '.curated'
-    # with open(curated_SYN_file, 'w') as fo, \
-    #         open(syri_out, 'r') as fi:
-    #     original_syri = fi.read()
-    #     fo.write(original_syri)
-    #     fo.write(formated_SYN)
+
+    #Do not wipe previous runs
+    putative_dir = "{}.{}.syri".format(qry_tail_fa, ref_tail_fa)
+    if not op.exists(putative_dir) or postprocess == "T":
+        qry_fadt = SeqIO.to_dict(SeqIO.parse(qry_fa, "fasta"))
+        ref_fadt = SeqIO.to_dict(SeqIO.parse(ref_fa, "fasta"))
+        with open(qry_tail_fa, 'w') as out_qry:
+            out_qry.write(qry_fadt[qry_chr][qry_offset - 1:].format('fasta'))
+        with open(ref_tail_fa, 'w') as out_ref:
+            out_ref.write(ref_fadt[ref_chr][ref_offset - 1:].format('fasta'))
+        syri(qry_tail_fa, ref_tail_fa, submit='T')
+    if postprocess == "T":
+        formated_SYN = format_syri_offset(qry_offset, ref_offset, "{}.{}.syri.out".format(qry_tail_fa, ref_tail_fa))
+        curated_SYN_file = syri_out + '.curated'
+        with open(curated_SYN_file, 'w') as fo, \
+            open(syri_out, 'r') as fi:
+            original_syri = fi.read()
+            fo.write(original_syri)
+            fo.write(formated_SYN)
     # synal_to_mosaic(curated_SYN_file, syri_out='T', output=curated_SYN_file + ".mosaic")
